Unresolved problem using all() showing up via test_3c284.py
First of all the range= keyword would hide the internal range() command, so I renamed it,
but my problem wasn't a side-effect of this. I did a bunch of stupid expansions to get to the bottom of
this, but it's persistently failing with TypeError: 'bool' object is not iterable

If I cut and paste the whole mess into an ipython session, it all works, the
original line
    if not (spectrum1.xarr.shape == spectrum2.xarr.shape) or not all(spectrum1.xarr == spectrum2.xarr):
just works. But running in the code, not. Is this one of those 1+1=3 cases?